WIP: unified implementation (3)
