+++ cdx_toolkit/filter_warc/args.py
@@ -51,13 +51,24 @@ def add_warcer_by_cdx_args(parser: argparse.ArgumentParser):
         '--parallel',
         type=int,
         default=1,
-        help='Number of parallel workers for fetchin WARC records (default: 1, sequential processing)',
+        help='Number of parallel workers for reading and writing WARC records (default: 1, sequential processing)',
+    )
+    parser.add_argument(
+        '--parallel_readers',
+        type=int,
+        default=None,
+        help='Number of parallel workers for reading WARC records (default: same as `parallel`)',
+    )
+    parser.add_argument(
+        '--parallel_writers',
+        type=int,
+        default=None,
+        help='Number of parallel workers for writing WARC records (default: same as `parallel`)',
     )
     parser.add_argument(
         '--log_every_n',
         type=int,
         default=1000,
         help='Every N extracted record a log message is emitted (0 = no record logs)',
-    )    
-    parser.add_argument('--implementation', type=str, default='fsspec', help='implementation (fsspec, aioboto3)')
+    )
     return parser
