+++ requirements.txt
@@ -6,6 +6,7 @@ warcio==1.7.4
 fsspec[s3]
 surt>=0.3.1
 tqdm>=4.67.1
+url-is-in>=0.1.1
 
 # used by Makefile
 pytest==6.2.4
